+++ src/main/java/li/cil/bedrockores/common/config/Constants.java
@@ -46,6 +46,7 @@ public final class Constants {
     static final String CONFIG_VEIN_CHANCE = "config.bedrockores.vein_chance";
     static final String CONFIG_VEIN_YIELD_BASE_SCALE = "config.bedrockores.vein_yield_base_scale";
     static final String CONFIG_VEIN_DISTANCE_SCALE_START = "config.bedrockores.vein_distance_scale_start";
+    static final String CONFIG_VEIN_MIN_SPAWN_DISTANCE = "config.bedrockores.vein_min_spawn_distance";
     static final String CONFIG_VEIN_DISTANCE_SCALE_MULTIPLIER = "config.bedrockores.vein_distance_scale_multiplier";
     static final String CONFIG_UI_ONLY_WHEN_SNEAKING = "config.bedrockores.ui_only_when_sneaking";
     static final String CONFIG_RETROGEN_SPEED = "config.bedrockores.retrogen_speed";