+++ changelog.md
@@ -1,1 +1,2 @@
+* Added setting to enforce minimum distance to spawn for veins to spawn.
 * Fixed potential crash when some mod adds a biome that doesn't have a registry name.